+++ modules/authcore/lib/Core/Workflow/Step/CreateAccountStep.php
@@ -1,13 +1,15 @@
 <?php
-
 /**
  * @author   Laurent Jouanneau
- * @copyright 2022-2023 Laurent Jouanneau
+ * @copyright 2022-2026 Laurent Jouanneau
  * @link     https://jelix.org
  * @license  MIT
  */
 
 namespace Jelix\Authentication\Core\Workflow\Step;
+
+use Jelix\Authentication\Core\Workflow\Event\CreateAccountEvent;
+use Jelix\Authentication\Core\Workflow\WorkflowState;
 
 class CreateAccountStep extends AbstractStep
 {
@@ -15,4 +17,16 @@ class CreateAccountStep extends AbstractStep
 
     protected $transition = 'account_created';
 
+    /**
+     * start the step.
+     *
+     * @param string $transition the name of the transition that is applied to start the step
+     * @return void
+     */
+    public function startStep($transition, WorkflowState $workflowState)
+    {
+        $event = new CreateAccountEvent($transition, $workflowState->getTemporaryUser(), $workflowState->getIdpId());
+        $this->eventDispatcher->dispatch($event);
+        $this->workflowState->setActions($event->getActions());
+    }
 }